move shared stuff to dmt/common
