+++ web/packages/plugins/apps/analysis-platform/src/App.tsx
@@ -3,8 +3,7 @@ import { Route } from 'react-router-dom'
 import { Layout } from 'antd'
 
 import Routes from './Routes'
-import { TApp, TLayout } from './Types'
-import { Header } from '@dmt/common'
+import { Header, TApp, TLayout } from '@dmt/common'
 import { backgroundColorDefault } from './components/Design/Colors'
 import Content from './components/Layout/Content'
 import Menu from './components/Layout/Menu'
