+++ web/packages/plugins/apps/analysis-platform/src/Enums.ts
@@ -1,15 +1,4 @@
-export enum Status {
-  STARTING = 'starting',
-  RUNNING = 'running',
-  FAILED = 'failed',
-  COMPLETED = 'completed',
-  UNKNOWN = 'unknown',
-}
-
-export enum Blueprints {
-  BLUEPRINT = 'system/SIMOS/Blueprint',
+export enum EBlueprints {
   ANALYSIS = 'AnalysisPlatformDS/Blueprints/Analysis',
-  BLOB = 'system/SIMOS/Blob',
   TASK = 'WorkflowDS/Blueprints/Task',
-  WORKFLOW = 'WorkflowDS/Blueprints/Workflow',
 }
