+++ web/packages/plugins/apps/analysis-platform/src/components/ErrorGroup.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+export const ErrorGroup = styled.div`
+  display: flex;
+  flex-direction: column;
+  border: 1px solid rgba(213, 18, 18, 0.71);
+  border-radius: 5px;
+  padding: 20px 20px;
+  background-color: #f6dfdf;
+`
