+++ web/packages/plugins/apps/analysis-platform/src/const.ts
@@ -1,9 +1,4 @@
 export const DEFAULT_DATASOURCE_ID = 'AnalysisPlatformDS'
-export const WORKFLOW_DATASOURCE = 'WorkflowDS'
-export const JOB = `${WORKFLOW_DATASOURCE}/Blueprints/Job`
-export const BLUEPRINTS = 'Blueprints'
 export const ENTITIES = 'Data'
 export const ANALYSIS_PATH = `${ENTITIES}/Analysis`
 export const ANALYSIS_RESULTS_PATH = `${DEFAULT_DATASOURCE_ID}/${ANALYSIS_PATH}/results`
-export const TASK = `${WORKFLOW_DATASOURCE}/${BLUEPRINTS}/Task`
-export const ANALYSIS = `${DEFAULT_DATASOURCE_ID}/${BLUEPRINTS}/Analysis`
