+++ web/packages/plugins/apps/analysis-platform/src/modules/View.tsx
@@ -1,20 +1,11 @@
 import React from 'react'
 // @ts-ignore
 import { useParams } from 'react-router-dom'
-import styled from 'styled-components'
 import { useDocument, UIPluginSelector } from '@dmt/common'
 // @ts-ignore
 import { NotificationManager } from 'react-notifications'
 import { TValidEntity } from '@dmt/common'
-
-export const ErrorGroup = styled.div`
-  display: flex;
-  flex-direction: column;
-  border: 1px solid rgba(213, 18, 18, 0.71);
-  border-radius: 5px;
-  padding: 20px 20px;
-  background-color: #f6dfdf;
-`
+import { ErrorGroup } from '../components/ErrorGroup'
 
 export const View = ({ settings }: any) => {
   const { data_source, entity_id } = useParams<{
