+++ web/packages/plugins/apps/analysis-platform/src/utils/auth.tsx
@@ -1,23 +1,3 @@
-import { getRoles } from '@dmt/common'
-
-const EXPERT_ROLES = [
-  'expert-operator',
-  'domain-developer',
-  'domain-expert',
-  'dmss-admin',
-]
-
-export function hasExpertRole(userData: any): boolean {
-  // Always return true if auth is not enabled
-  if (process.env.REACT_APP_AUTH !== '1') return true
-  const roles = getRoles(userData)
-  if (roles) {
-    // If any of the roles the user has in the the EXPERT_ROLES array, return TRUE
-    return roles.some((role: string) => EXPERT_ROLES.includes(role))
-  }
-  return false
-}
-
 export function getUsername(userData: any): string {
   if (userData?.preferred_username) {
     // With azuare AD, the user's email is stored in the preferred_username attribute.
