+++ web/packages/plugins/shared/common/src/utils/variables.tsx
@@ -1,11 +1,3 @@
-export enum BlueprintEnum {
-  BLUEPRINT = 'system/SIMOS/Blueprint',
-  ATTRIBUTE = 'system/SIMOS/BlueprintAttribute',
-  PACKAGE = 'system/SIMOS/Package',
-  ENTITY = 'system/SIMOS/Entity',
-  ENUM = 'system/SIMOS/Enum',
-}
-
 export const PATH_INPUT_FIELD_WIDTH: string = '450px'
 export const INPUT_FIELD_WIDTH: string = '285px'
 export const TREE_DIALOG_HEIGHT: string = '50vh'
